refactor: 배치용 내부 DTO 패키지 분리 - #76
