+++ settlement/src/main/java/com/back/settlement/app/dto/internal/SettlementWithItems.java
@@ -1,4 +1,4 @@
-package com.back.settlement.app.dto.request;
+package com.back.settlement.app.dto.internal;
 
 import com.back.settlement.domain.Settlement;
 import com.back.settlement.domain.SettlementItem;
